LayoutTests/imported/w3c/web-platform-tests/css/css-grid/masonry/tentative/masonry-parsing-expected.txt


FAIL e.style['grid-template-columns'] = "masonry" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['grid-template-columns'] = "masonry []" should not set the property value
PASS e.style['grid-template-columns'] = "masonry [foo] 40px" should not set the property value
PASS e.style['grid-template-columns'] = "masonry 40px" should not set the property value
PASS e.style['grid-template-columns'] = "[foo] masonry" should not set the property value
PASS e.style['grid-template-columns'] = "0px masonry" should not set the property value
PASS e.style['grid-template-columns'] = "masonry masonry" should not set the property value
PASS e.style['grid-template-columns'] = "subgrid masonry" should not set the property value
PASS e.style['grid-template-columns'] = "masonry subgrid" should not set the property value
PASS e.style['grid-template-columns'] = "masonry repeat(1, [])" should not set the property value
PASS e.style['grid-template-columns'] = "masonry repeat(auto-fill, [])" should not set the property value
PASS e.style['grid-template-columns'] = "masonry / none" should not set the property value
FAIL e.style['grid-template-rows'] = "masonry" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['grid-template-rows'] = "masonry []" should not set the property value
PASS e.style['grid-template-rows'] = "masonry [foo] 40px" should not set the property value
PASS e.style['grid-template-rows'] = "masonry 40px" should not set the property value
PASS e.style['grid-template-rows'] = "[foo] masonry" should not set the property value
PASS e.style['grid-template-rows'] = "0px masonry" should not set the property value
PASS e.style['grid-template-rows'] = "masonry masonry" should not set the property value
PASS e.style['grid-template-rows'] = "subgrid masonry" should not set the property value
PASS e.style['grid-template-rows'] = "masonry subgrid" should not set the property value
PASS e.style['grid-template-rows'] = "masonry repeat(1, [])" should not set the property value
PASS e.style['grid-template-rows'] = "masonry repeat(auto-fill, [])" should not set the property value
PASS e.style['grid-template-rows'] = "masonry / none" should not set the property value
FAIL e.style['grid-template'] = "masonry / masonry" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['grid-template'] = "masonry / masonry" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "masonry" but got ""
FAIL e.style['grid-template'] = "masonry / masonry" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "masonry" but got ""
FAIL e.style['grid-template'] = "masonry / masonry" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid-template'] = "masonry / subgrid [foo]" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['grid-template'] = "masonry / subgrid [foo]" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "subgrid [foo]" but got ""
FAIL e.style['grid-template'] = "masonry / subgrid [foo]" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "masonry" but got ""
FAIL e.style['grid-template'] = "masonry / subgrid [foo]" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid-template'] = "10px / masonry" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['grid-template'] = "10px / masonry" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "masonry" but got ""
FAIL e.style['grid-template'] = "10px / masonry" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "10px" but got ""
FAIL e.style['grid-template'] = "10px / masonry" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid'] = "masonry / masonry" should set grid-auto-columns assert_equals: grid-auto-columns should be canonical expected "auto" but got ""
FAIL e.style['grid'] = "masonry / masonry" should set grid-auto-flow assert_equals: grid-auto-flow should be canonical expected "row" but got ""
FAIL e.style['grid'] = "masonry / masonry" should set grid-auto-rows assert_equals: grid-auto-rows should be canonical expected "auto" but got ""
FAIL e.style['grid'] = "masonry / masonry" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['grid'] = "masonry / masonry" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "masonry" but got ""
FAIL e.style['grid'] = "masonry / masonry" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "masonry" but got ""
FAIL e.style['grid'] = "masonry / masonry" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid'] = "masonry / subgrid [foo]" should set grid-auto-columns assert_equals: grid-auto-columns should be canonical expected "auto" but got ""
FAIL e.style['grid'] = "masonry / subgrid [foo]" should set grid-auto-flow assert_equals: grid-auto-flow should be canonical expected "row" but got ""
FAIL e.style['grid'] = "masonry / subgrid [foo]" should set grid-auto-rows assert_equals: grid-auto-rows should be canonical expected "auto" but got ""
FAIL e.style['grid'] = "masonry / subgrid [foo]" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['grid'] = "masonry / subgrid [foo]" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "subgrid [foo]" but got ""
FAIL e.style['grid'] = "masonry / subgrid [foo]" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "masonry" but got ""
FAIL e.style['grid'] = "masonry / subgrid [foo]" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['grid'] = "10px / masonry" should set grid-auto-columns assert_equals: grid-auto-columns should be canonical expected "auto" but got ""
FAIL e.style['grid'] = "10px / masonry" should set grid-auto-flow assert_equals: grid-auto-flow should be canonical expected "row" but got ""
FAIL e.style['grid'] = "10px / masonry" should set grid-auto-rows assert_equals: grid-auto-rows should be canonical expected "auto" but got ""
FAIL e.style['grid'] = "10px / masonry" should set grid-template-areas assert_equals: grid-template-areas should be canonical expected "none" but got ""
FAIL e.style['grid'] = "10px / masonry" should set grid-template-columns assert_equals: grid-template-columns should be canonical expected "masonry" but got ""
FAIL e.style['grid'] = "10px / masonry" should set grid-template-rows assert_equals: grid-template-rows should be canonical expected "10px" but got ""
FAIL e.style['grid'] = "10px / masonry" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['masonry-auto-flow'] = "pack" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['masonry-auto-flow'] = "pack ordered" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['masonry-auto-flow'] = "ordered next" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['masonry-auto-flow'] = "next definite-first" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['masonry-auto-flow'] = "definite-first pack" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['masonry-auto-flow'] = "auto" should not set the property value
PASS e.style['masonry-auto-flow'] = "none" should not set the property value
PASS e.style['masonry-auto-flow'] = "10px" should not set the property value
PASS e.style['masonry-auto-flow'] = "row" should not set the property value
PASS e.style['masonry-auto-flow'] = "dense" should not set the property value
